services/generator/requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python
python-dotenv